moved visuals.jl in a separate directory called plot_utils
